Moving refactoring changes to this earlier MR
