+++ roslibrust/src/rosbridge/mod.rs
@@ -30,46 +30,13 @@ mod topic_provider;
 mod comm;
 
 use futures_util::stream::{SplitSink, SplitStream};
-use log::*;
 use std::collections::HashMap;
 use tokio::net::TcpStream;
 use tokio_tungstenite::*;
 use tungstenite::Message;
 
-/// For now starting with a central error type, may break this up more in future
-#[derive(thiserror::Error, Debug)]
-pub enum RosLibRustError {
-    #[error("Not currently connected to ros master / bridge")]
-    Disconnected,
-    // TODO we probably want to eliminate tungstenite from this and hide our
-    // underlying websocket implementation from the API
-    // currently we "technically" break the API when we change tungstenite verisons
-    #[error("Websocket communication error: {0}")]
-    CommFailure(tokio_tungstenite::tungstenite::Error),
-    #[error("Operation timed out: {0}")]
-    Timeout(#[from] tokio::time::error::Elapsed),
-    #[error("Failed to parse message from JSON: {0}")]
-    InvalidMessage(#[from] serde_json::Error),
-    #[error("Rosbridge server reported an error: {0}")]
-    ServerError(String),
-    // Generic catch-all error type for not-yet-handled errors
-    // TODO ultimately this type will be removed from API of library
-    #[error(transparent)]
-    Unexpected(#[from] anyhow::Error),
-}
-
-/// Provides an implementation tranlating the underlying websocket error into our error type
-impl From<tokio_tungstenite::tungstenite::Error> for RosLibRustError {
-    fn from(e: tokio_tungstenite::tungstenite::Error) -> Self {
-        // TODO we probably want to expand this type and do some matching here
-        RosLibRustError::CommFailure(e)
-    }
-}
-
-/// Generic result type used as standard throughout library.
-/// Note: many functions which currently return this will be updated to provide specific error
-/// types in the future instead of the generic error here.
-pub type RosLibRustResult<T> = Result<T, RosLibRustError>;
+// Doing this to maintain backwards compatibilities like `use roslibrust::rosbridge::RosLibRustError`
+pub use super::{RosLibRustError, RosLibRustResult};
 
 /// Used for type erasure of message type so that we can store arbitrary handles
 type Callback = Box<dyn Fn(&str) + Send + Sync>;
